Add files via upload
Version 2: still only a single two point neuron, but moving in the direction of creting a full network. runTPN_spines_shun_v2 sets the paramaters for calling TPN_spines_shun_v2 which in turn calls TPN_runstepfor each time step.It is notefficient, but aims to record depolarisation at each place in the neuron so that one can examione exactly what is happening.